requirements.txt
fastapi
uvicorn
tensorflow
pydantic
numpy
Pillow
